Fix temperature type

+++ panther_hardware_interfaces/test/include/mock_roboteq.hpp
@@ -57,7 +57,7 @@ public:
   void SetDriverFaultFlag(DriverFaultFlags flag);
   void SetDriverScriptFlag(DriverScriptFlags flag);
   void SetDriverRuntimeError(uint8_t channel, DriverRuntimeErrors flag);
-  void SetTemperature(int16_t value) { (*this)[0x210F][1] = value; }
+  void SetTemperature(int8_t value) { (*this)[0x210F][1] = value; }
   void SetVoltage(uint16_t value) { (*this)[0x210D][2] = value; }
   void SetBatAmps1(int16_t value) { (*this)[0x210C][1] = value; }
   void SetBatAmps2(int16_t value) { (*this)[0x210C][2] = value; }
